Removal of Localized functions
I am currently looking into the tracking of addons that enable and disable Combat Logging sadly the localization of the functions make it impossible to be hooked for a addon or script that loads after loggerhead. The optimization gain of the localization is minimal anyway. I also removed the other localized functions where some werent even used.

I hope you can accept the PR and this wasnt to bad.